+++ podaac/subsetter/subset.py
@@ -354,7 +354,8 @@ def subset(file_to_subset: str, bbox: np.ndarray, output_file: str,
                     time_encoding[group_path][var_name]['calendar'] = calendar
                 if units:
                     time_encoding[group_path][var_name]['units'] = units
-                time_encoding[group_path][var_name]['dtype'] = dtype
+                if dtype and units:
+                    time_encoding[group_path][var_name]['dtype'] = dtype
                 if calendar:
                     time_calendar_attributes[time] = calendar